requirements.txt
absl-py==1.2.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
bio==1.3.9
biopython==1.79
biothings-client==0.2.6
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.1.0
cirq-core==0.14.1
cirq-google==0.14.1
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
dill==0.3.5.1
duet==0.2.7
entrypoints==0.4
etils==0.6.0
executing==0.9.1
filelock==3.7.1
flatbuffers==2.0
fonttools==4.34.4
gast==0.4.0
google-api-core==1.21.0
google-auth==1.18.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.48.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
intel_tensorflow==2.7.0
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mpmath==1.2.1
mygene==3.2.2
nest-asyncio==1.5.5
networkx==2.8.5
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.17.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.25
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.3.0
sympy==1.8
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.9.0
tensorflow-quantum==0.6.1
termcolor==1.1.0
tokenizers==0.12.1
toml==0.10.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.0
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
Werkzeug==2.2.1
wrapt==1.14.1
zipp==3.8.1
